chore(validator): move Validator and ValidatorTest from policy to validator package for structural consistency
- Relocated Validator.kt to validator package to better reflect its responsibility
- Moved corresponding ValidatorTest to match the new structure
- Cleaned up unused policy references
