+++ src/test/kotlin/lotto/validator/ValidatorTest.kt
@@ -1,9 +1,10 @@
-package lotto.policy
+package lotto.validator
 
 import lotto.error.Bonus
 import lotto.error.Purchase
 import lotto.error.Winning
-import org.assertj.core.api.Assertions.assertThat
+import lotto.policy.GamePolicy
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 
 class ValidatorTest {
@@ -11,42 +12,42 @@ class ValidatorTest {
 	fun `returns INVALID_SIZE error when lotto numbers have more than 6 elements`() {
 		val invalidSizeWinning = listOf(1, 2, 3, 4, 5, 6, 7)
 		val errorType = Validator.isValidWinningNumbers(invalidSizeWinning)
-		assertThat(errorType).isEqualTo(Winning.INVALID_SIZE)
+		Assertions.assertThat(errorType).isEqualTo(Winning.INVALID_SIZE)
 	}
 
 	@Test
 	fun `returns INVALID_SIZE error when lotto numbers have less than 6 elements`() {
 		val invalidSizeWinning = listOf(1, 2, 3, 4, 5)
 		val errorType = Validator.isValidWinningNumbers(invalidSizeWinning)
-		assertThat(errorType).isEqualTo(Winning.INVALID_SIZE)
+		Assertions.assertThat(errorType).isEqualTo(Winning.INVALID_SIZE)
 	}
 
 	@Test
 	fun `returns null when lotto numbers are valid`() {
 		val winning = listOf(1, 2, 3, 4, 5, 6)
 		val errorType = Validator.isValidWinningNumbers(winning)
-		assertThat(errorType).isEqualTo(null)
+		Assertions.assertThat(errorType).isEqualTo(null)
 	}
 
 	@Test
 	fun `returns DUPLICATE_NUMBER error when lotto numbers contain duplicates`() {
 		val duplicatesWinning = listOf(1, 2, 3, 4, 5, 5)
 		val errorType = Validator.isValidWinningNumbers(duplicatesWinning)
-		assertThat(errorType).isEqualTo(Winning.DUPLICATE_NUMBER)
+		Assertions.assertThat(errorType).isEqualTo(Winning.DUPLICATE_NUMBER)
 	}
 
 	@Test
 	fun `returns NOT_IN_RANGE error when lotto numbers are not in range (too small)`() {
 		val outOfRangeWinning = listOf(-1, 0, 1, 2, 3, 4)
 		val errorType = Validator.isValidWinningNumbers(outOfRangeWinning)
-		assertThat(errorType).isEqualTo(Winning.NOT_IN_RANGE)
+		Assertions.assertThat(errorType).isEqualTo(Winning.NOT_IN_RANGE)
 	}
 
 	@Test
 	fun `returns NOT_IN_RANGE error when lotto numbers are not in range (too big)`() {
 		val outOfRangeWinning = listOf(2021, 401, 1, 2, 3, 4)
 		val errorType = Validator.isValidWinningNumbers(outOfRangeWinning)
-		assertThat(errorType).isEqualTo(Winning.NOT_IN_RANGE)
+		Assertions.assertThat(errorType).isEqualTo(Winning.NOT_IN_RANGE)
 	}
 
 	@Test
@@ -54,7 +55,7 @@ class ValidatorTest {
 		val winning = listOf(1, 2, 3, 4, 5, 6)
 		val bonus = winning[0]
 		val errorType = Validator.isValidBonusNumber(bonus, winning)
-		assertThat(errorType).isEqualTo(Bonus.DUPLICATE_NUMBER)
+		Assertions.assertThat(errorType).isEqualTo(Bonus.DUPLICATE_NUMBER)
 	}
 
 	@Test
@@ -62,7 +63,7 @@ class ValidatorTest {
 		val winning = listOf(1, 2, 3, 4, 5, 6)
 		val bonus = 4242
 		val errorType = Validator.isValidBonusNumber(bonus, winning)
-		assertThat(errorType).isEqualTo(Bonus.NOT_IN_RANGE)
+		Assertions.assertThat(errorType).isEqualTo(Bonus.NOT_IN_RANGE)
 	}
 
 
@@ -71,7 +72,7 @@ class ValidatorTest {
 		val winning = listOf(1, 2, 3, 4, 5, 6)
 		val bonus = winning[0]
 		val errorType = Validator.isValidBonusNumber(bonus, winning)
-		assertThat(errorType).isEqualTo(Bonus.DUPLICATE_NUMBER)
+		Assertions.assertThat(errorType).isEqualTo(Bonus.DUPLICATE_NUMBER)
 	}
 
 
@@ -79,13 +80,13 @@ class ValidatorTest {
 	fun `should return NOT_DIVISIBLE_BY_TICKET_PRICE error when amount is not multiple of ticket price`() {
 		val amountNotDivisibleByTicketPrice = GamePolicy.TICKET_PRICE * 123 + 1L
 		val errorType = Validator.isValidAmount(amountNotDivisibleByTicketPrice)
-		assertThat(errorType).isEqualTo(Purchase.NOT_DIVISIBLE_BY_TICKET_PRICE)
+		Assertions.assertThat(errorType).isEqualTo(Purchase.NOT_DIVISIBLE_BY_TICKET_PRICE)
 	}
 
 	@Test
 	fun `should return INVALID_PURCHASE_AMOUNT error when amount is not multiple of ticket price`() {
 		val amountNotPurchase = GamePolicy.TICKET_PRICE - 42L
 		val errorType = Validator.isValidAmount(amountNotPurchase)
-		assertThat(errorType).isEqualTo(Purchase.INVALID_PURCHASE_AMOUNT)
+		Assertions.assertThat(errorType).isEqualTo(Purchase.INVALID_PURCHASE_AMOUNT)
 	}
 }
